+++ test/utils/df-config-browser.js
@@ -1,0 +1,4 @@
+module.exports = {
+  remote: true,
+  port: 25000
+}
