+++ test/utils/df-config-nodejs.js
@@ -1,0 +1,3 @@
+module.exports = {
+  remote: false
+}
